backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9 